Refactor: Convert Login component from JSX to TypeScript (tsx)
