+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "@emotion/react": "^11.11.3",
+    "@fortawesome/free-brands-svg-icons": "^6.5.1",
     "@fortawesome/free-solid-svg-icons": "^6.5.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@tanstack/react-query": "^5.17.9",
